requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
pg8000==1.29.8
APScheduler==3.10.4
python-dateutil==2.8.2
